Configuring dotenv:
The dotenv.config() function is called with a configuration object. The path property specifies the location of the .env file. In this case, it's set to "../config/.env", which means the file is located one directory up from the current file, inside a config folder. This ensures that the environment variables defined in the .env file are loaded into process.env.