Schema Fields:
name:
Type: String
Required: true
This field stores the user's name and is mandatory.
username:
Type: String
Required: true
Unique: true
This field stores the user's unique username, which must be provided and must be unique across all documents in the collection.
email:
Type: String
Required: true
Unique: true
This field stores the user's unique email address, which must be provided and must be unique across all documents in the collection.
password:
Type: String
Required: true
This field stores the user's password and is mandatory. Typically, passwords are stored as hashed values for security reasons, though this is not handled directly in the schema.
followers:
Type: Array
Default: []
This field stores an array of user IDs representing the followers of the user. By default, it is an empty array.
following:
Type: Array
Default: []
This field stores an array of user IDs representing the users that this user is following. By default, it is an empty array.
bookmarks:
Type: Array
Default: []
This field stores an array of bookmarked items (could be posts, articles, etc.). By default, it is an empty array.