Creating the Model:
The mongoose.model() function is used to create a model from the schema. The first argument is the name of the model ("User"), and the second argument is the schema (userSchema). The resulting User model is a constructor function that can be used to create and query documents in the User collection in the database.