getMyProfile Function Defining the getMyProfile Function:
This function retrieves the profile of the logged-in user. It is asynchronous because it interacts with the database.
Extracting Data from the Request:

id is extracted from req.params.
Fetching the User Profile:

Uses User.findById(id).select("-password") to find the user by ID and exclude the password field from the response.
Sending Success Response:

Returns a JSON response with the user's profile.