getOtherUsers Function Defining the getOtherUsers Function:
This function retrieves profiles of all users except the logged-in user. It is asynchronous because it interacts with the database.
Extracting Data from the Request:

id is extracted from req.params.
Fetching Other Users:

Uses User.find({ _id: { $ne: id } }).select("-password") to find all users except the logged-in user and exclude the password field from the response.
Sending Success Response:

Returns a JSON response with the profiles of the other users.