follow Function Defining the follow Function:
This function handles following a user. It is asynchronous because it interacts with the database.
Extracting Data from the Request:

loggedInUserId and userId are extracted from req.body and req.params, respectively.
Fetching Users:

Finds the logged-in user and the user to be followed by their IDs.
Checking and Adding Follower:

Checks if the logged-in user is already following the user. If not, it adds the logged-in user to the user's followers and adds the user to the logged-in user's following.
Sending Success Response:

Returns a JSON response indicating successful follow.