unfollow Function Defining the unfollow Function:
This function handles unfollowing a user. It is asynchronous because it interacts with the database.
Extracting Data from the Request:

loggedInUserId and userId are extracted from req.body and req.params, respectively.
Fetching Users:

Finds the logged-in user and the user to be unfollowed by their IDs.
Checking and Removing Follower:

Checks if the logged-in user is following the user. If so, it removes the logged-in user from the user's followers and removes the user from the logged-in user's following.
Sending Success Response:

Returns a JSON response indicating successful unfollow.