Configuration and Initialization Loading Environment Variables:
dotenv.config({ path: ".env" });
This loads the environment variables from the .env file located in the current directory, making them available through process.env.
Connecting to the Database:

databaseConnection();
This invokes the databaseConnection function to establish a connection to the MongoDB database using Mongoose.
Creating the Express Application:

const app = express();
This creates an instance of an Express application, which will be used to define routes and middleware.