Reducers Reducers: The reducers object defines various reducer functions that handle different actions. Each function updates the state based on the action's payload.
getUser:

getUser: (state, action) => { state.user = action.payload; },
This reducer sets the user state to the value provided in the action's payload.
getOtherUsers:

getOtherUsers: (state, action) => { state.otherUsers = action.payload; },
This reducer sets the otherUsers state to the value provided in the action's payload.
getMyProfile:

getMyProfile: (state, action) => { state.profile = action.payload; },
This reducer sets the profile state to the value provided in the action's payload.
followingUpdate:

followingUpdate: (state, action) => { ... },
This reducer updates the following array of the user state. It checks if the user is already following the user ID provided in the action's payload:
If the user is following the ID, it removes the ID from the following array (unfollow).
If the user is not following the ID, it adds the ID to the following array (follow).