Initial State:
initialState: { tweets: null, refresh: false, isActive: true },
Defines the initial state of the slice with three properties:
tweets: Initially null, this will hold an array of tweets or null.
refresh: Initially false, used as a boolean flag to trigger a refresh or re-rendering of tweets.
isActive: Initially true, indicates the active state of the component or feature related to tweets.