Reducers Reducers: The reducers object defines various reducer functions that handle state updates based on dispatched actions.
getAllTweets:

getAllTweets: (state, action) => { state.tweets = action.payload; },
This reducer sets the tweets state to the value provided in the action's payload. It typically updates the state with an array of tweets fetched from an API or received from another part of the application.
getRefresh:

getRefresh: (state) => { state.refresh = !state.refresh; },
This reducer toggles the refresh state. It flips the boolean value of refresh, which can be useful for triggering a re-render or refresh action in the UI.
getIsActive:

getIsActive: (state, action) => { state.isActive = action.payload; },
This reducer sets the isActive state based on the payload received in the action. It updates whether the component or feature related to tweets is active or not.