JSX Structure:
<div className="App"> ... </div>
The App component returns JSX, which represents the structure of the UI to be rendered.
className="App" applies the CSS class App defined in App.css to the root div of the component.
Rendering Child Components:

<Body />
Renders the Body component imported earlier. This suggests that the main content or body of the application is encapsulated within the Body component.
<Toaster />
Renders the Toaster component from react-hot-toast. This component is used for displaying toast notifications in response to various actions or events within the application.