Event Handlers (likeOrDislikeHandler, deleteTweetHandler):
likeOrDislikeHandler: Handles the liking or disliking of a tweet.

Sends a PUT request to the server using axios.put to toggle the like status of the tweet identified by tweet._id.
Dispatches getRefresh() to update the tweet list after liking or disliking.
Shows a success toast notification if the operation is successful; otherwise, shows an error toast.